fetched last messages on the home screen
