+++ pubspec.yaml
@@ -39,6 +39,8 @@ dependencies:
   firebase_auth: ^5.4.1
   intl: ^0.20.2
   provider: ^6.1.2
+  flutter_screenutil: ^5.9.3
+  timeago: ^3.2.2
 
 dev_dependencies:
   flutter_test:
